Enables code coverage reporting (#17)
* Enables code coverage reporting

Adds code coverage reporting to the project using grcov and uploads the results to Codecov.

This change introduces:
- Integration with Codecov for coverage reporting.
- Makefile targets for generating and viewing coverage reports locally.
- GitHub Actions configuration to run coverage in CI.
- Ignore coverage artifacts in .gitignore.

The setup-coverage script automates the installation of necessary tools and configures the environment.

* Generates coverage badge and updates README

Adds functionality to generate a coverage badge based on test coverage reports.

The coverage percentage is extracted from the generated reports, and a badge URL is created with a color-coded indicator of the coverage level. The badge URL and the downloaded badge are saved for inclusion in the README.

Updates the README with instructions on how to generate the coverage badge.
Removes the dependency on Codecov.