Refactors test setup and adds LocalStack config (#18)
* Refactors test setup and adds LocalStack config

Moves Docker Compose files to a dedicated "docker" directory for better organization.

Updates the Makefile to reference the new Docker Compose file locations.

Adds default AWS configuration to `.env.example` for LocalStack.

Changes the default Postgres port in the test environment to avoid conflicts.

Updates the README to reflect the changes in project structure and setup instructions.

* Updates Docker Compose paths and script paths

Updates references to docker-compose files and shell scripts to use the `docker/` and `scripts/` directories, respectively. This ensures consistency and correct execution of commands within the project and CI environment.

Adds checks within `setup-localstack.sh` to ensure LocalStack is running and SQS is enabled before proceeding.

* Fixes postgres test port mapping

Corrects the port mapping in the test docker-compose file
to use the standard Postgres port 5432.
This ensures the test database is accessible on the correct port.